Add merge only columns to notion
Implement merge only columns option to Notion models and upload
